+++ lib/assets/app_images.dart
@@ -1,0 +1,5 @@
+class AppImages {
+  static const String flagBR = 'assets/flags/brazil.svg';
+  static const String flagUS = 'assets/flags/usa.svg';
+  static const String flagES = 'assets/flags/spain.svg';
+}