+++ lib/mod_layout_one.dart
@@ -49,6 +49,7 @@ export 'widgets/grid_system/grid_system_container.dart';
 export 'widgets/grid_system/grid_system_rows.dart';
 export 'widgets/modal/modal.dart';
 export 'widgets/tabs/tabs.dart';
+export 'widgets/text/text_divider.dart';
 export 'widgets/textbox/textbox.dart';
 
 // Configuration