+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## 1.0.8
+
+* Correction of language translation, set to English.
+
+## 1.0.7
+
+* Updated LanguageSelector widget to use PopupMenuButton for language selection.
+* Improved language flag display in LanguageSelector widget.
+
+## 1.0.6
+
+* Added footerHeight parameter to ModBaseLayout to allow customization of footer height.
+
+## 1.0.5
+
+* Added onTap parameter to MenuItem in sidebar.
+
+## 1.0.4
+
+* Added new examples to ButtonsPage.
+* Improved documentation in README.md.
+
+## 1.0.3
+
+* Added ModTextDivider widget.
+
+## 1.0.2
+
+* Improvement in the button.
+
+## 1.0.1
+
+* Initial release with basic features.
+
 ## 1.0.0
 
-* TODO: Describe initial release.
+* Initial release with basic features.